Includes: Refactor verilated.h and deprecate verilated_heavy.h (#2701).
